// src/config/intefaces/config.interface.ts
export interface ConfigInterface {
  NODE_ENV: string;
  SERVER_PORT: number;
  MONGO_URI: string;
  JWT_SECRET: string;
  AWS_ACCESS_KEY: string;
  AWS_SECRET_ACCESS_KEY: string;
  AWS_BUCKET_NAME: string;
  AWS_S3_REGION: string;
  AWS_SES_REGION: string;
  AWS_SNS_REGION: string;
  AWS_BUCKET_URI: string;
  AWS_MAX_FILE_UPLOAD_SIZE_MB: number;
  AWS_MAX_FILES_UPLOAD_ALLOW: number;
  REDIS_URI: string;
  SLACK_AUTHORIZATION_URI: string;
  SLACK_CLIENT_ID: string;
  SLACK_CLIENT_SECRET: string;
  SLACK_ALLOWED_REDIRECT_URI_LIST: string[];
  SLACK_SCOPES: string;
  GOOGLE_WEB_CLIENT_ID: string;
  GOOGLE_WEB_CLIENT_SECRET: string;
  GOOGLE_API_KEY: string;
  GOOGLE_WEB_CALL_BACK_URL: string;
  GOOGLE_IOS_CLIENT_ID: string;
  GOOGLE_ANDROID_CLIENT_ID: string;
  MONGO_URI_TEST: string;
  MAILGUN_API_KEY: string;
  FRONTEND_APP_LINK: string;
  ONESIGNAL_APP_ID: string;
  ONESIGNAL_REST_API_KEY: string;
  SLACK_BASE_URL: string;
  TRELLO_AUTHORIZATION_URI: string;
  TRELLO_TOKEN_EXPIRY: string;
  TRELLO_APP_NAME: string;
  TRELLO_SCOPES: string;
  TRELLO_RESPONSE_TYPE: string;
  TRELLO_API_KEY: string;
  API_URL: string;
  GITHUB_AUTHORIZATION_URI: string;
  GITHUB_CLIENT: string;
  GITHUB_SECRET: string;
  GITHUB_REDIRECT_URL: string;
  FACEBOOK_APP_ID: string;
  FACEBOOK_APP_SECRET: string;
  FACEBOOK_CALLBACK_URL: string;
  FIREBASE_PROJECT_ID: string;
  FIREBASE_PRIVATE_KEY: string;
  FIREBASE_CLIENT_EMAIL: string;
  ELASTICSEARCH_HOST: string;
  ELASTICSEARCH_HOST_URL: string;
  ELASTIC_SEARCH_MICRO_SERVICE_PORT: number;
  MONGO_OP_LOG_MICRO_SERVICE_PORT: number;
  ORDER_MICRO_SERVICE_PORT: number;
  KAFKA_PORT: number;
}